requirements.txt
alembic==1.4.3
aniso8601==8.1.0
bcrypt==3.1.7
cffi==1.14.4
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.21.0
permissive-dict==1.0.3
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.2.0
